The new device template for generic firewire camera. Again with a better sql2cpp

+++ src/pluto_main/Table_CachedScreens.cpp
@@ -326,7 +326,7 @@ if (is_null[3])
 return "NULL";
 
 char *buf = new char[39];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Modification_LastGen.c_str(), (unsigned long) min(19,m_Modification_LastGen.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Modification_LastGen.c_str(), (unsigned long) min((size_t)19,m_Modification_LastGen.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -418,7 +418,7 @@ if (is_null[10])
 return "NULL";
 
 char *buf = new char[29];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min(14,m_psc_mod.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min((size_t)14,m_psc_mod.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
