+++ src/pluto_main/Table_ConfigType_File.cpp
@@ -291,7 +291,7 @@ if (is_null[2])
 return "NULL";
 
 char *buf = new char[131071];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_InputFile.c_str(), (unsigned long) min(65535,m_InputFile.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_InputFile.c_str(), (unsigned long) min((size_t)65535,m_InputFile.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -305,7 +305,7 @@ if (is_null[3])
 return "NULL";
 
 char *buf = new char[131071];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_OutputFile.c_str(), (unsigned long) min(65535,m_OutputFile.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_OutputFile.c_str(), (unsigned long) min((size_t)65535,m_OutputFile.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -384,7 +384,7 @@ if (is_null[9])
 return "NULL";
 
 char *buf = new char[29];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min(14,m_psc_mod.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min((size_t)14,m_psc_mod.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
