+++ src/pluto_main/Table_ConfigType_Token.cpp
@@ -283,7 +283,7 @@ if (is_null[2])
 return "NULL";
 
 char *buf = new char[61];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_FindStr.c_str(), (unsigned long) min(30,m_FindStr.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_FindStr.c_str(), (unsigned long) min((size_t)30,m_FindStr.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -297,7 +297,7 @@ if (is_null[3])
 return "NULL";
 
 char *buf = new char[131071];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_ReplaceStr.c_str(), (unsigned long) min(65535,m_ReplaceStr.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_ReplaceStr.c_str(), (unsigned long) min((size_t)65535,m_ReplaceStr.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -363,7 +363,7 @@ if (is_null[8])
 return "NULL";
 
 char *buf = new char[29];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min(14,m_psc_mod.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min((size_t)14,m_psc_mod.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
