+++ src/pluto_main/Table_DHCPDevice_DeviceData.cpp
@@ -281,7 +281,7 @@ if (is_null[2])
 return "NULL";
 
 char *buf = new char[131071];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_IK_DeviceData.c_str(), (unsigned long) min(65535,m_IK_DeviceData.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_IK_DeviceData.c_str(), (unsigned long) min((size_t)65535,m_IK_DeviceData.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -347,7 +347,7 @@ if (is_null[7])
 return "NULL";
 
 char *buf = new char[29];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min(14,m_psc_mod.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min((size_t)14,m_psc_mod.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
