+++ src/pluto_main/Table_DeviceTemplate_AV.cpp
@@ -414,7 +414,7 @@ if (is_null[8])
 return "NULL";
 
 char *buf = new char[9];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_NumericEntry.c_str(), (unsigned long) min(4,m_NumericEntry.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_NumericEntry.c_str(), (unsigned long) min((size_t)4,m_NumericEntry.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -493,7 +493,7 @@ if (is_null[14])
 return "NULL";
 
 char *buf = new char[29];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min(14,m_psc_mod.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min((size_t)14,m_psc_mod.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
