+++ src/pluto_main/Table_DeviceTemplate_DeviceData.cpp
@@ -363,7 +363,7 @@ if (is_null[2])
 return "NULL";
 
 char *buf = new char[131071];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_IK_DeviceData.c_str(), (unsigned long) min(65535,m_IK_DeviceData.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_IK_DeviceData.c_str(), (unsigned long) min((size_t)65535,m_IK_DeviceData.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -416,7 +416,7 @@ if (is_null[6])
 return "NULL";
 
 char *buf = new char[131071];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Description.c_str(), (unsigned long) min(65535,m_Description.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Description.c_str(), (unsigned long) min((size_t)65535,m_Description.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -456,7 +456,7 @@ if (is_null[9])
 return "NULL";
 
 char *buf = new char[51];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_ShortDescription.c_str(), (unsigned long) min(25,m_ShortDescription.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_ShortDescription.c_str(), (unsigned long) min((size_t)25,m_ShortDescription.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -522,7 +522,7 @@ if (is_null[14])
 return "NULL";
 
 char *buf = new char[29];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min(14,m_psc_mod.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min((size_t)14,m_psc_mod.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
