+++ src/pluto_main/Table_DeviceTemplate_MediaType.cpp
@@ -408,7 +408,7 @@ if (is_null[8])
 return "NULL";
 
 char *buf = new char[81];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Extensions.c_str(), (unsigned long) min(40,m_Extensions.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Extensions.c_str(), (unsigned long) min((size_t)40,m_Extensions.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -474,7 +474,7 @@ if (is_null[13])
 return "NULL";
 
 char *buf = new char[29];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min(14,m_psc_mod.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min((size_t)14,m_psc_mod.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
