+++ src/pluto_main/Table_Device_MRU.cpp
@@ -294,7 +294,7 @@ if (is_null[2])
 return "NULL";
 
 char *buf = new char[511];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Filename.c_str(), (unsigned long) min(255,m_Filename.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Filename.c_str(), (unsigned long) min((size_t)255,m_Filename.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -308,7 +308,7 @@ if (is_null[3])
 return "NULL";
 
 char *buf = new char[39];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_LastUsed.c_str(), (unsigned long) min(19,m_LastUsed.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_LastUsed.c_str(), (unsigned long) min((size_t)19,m_LastUsed.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -374,7 +374,7 @@ if (is_null[8])
 return "NULL";
 
 char *buf = new char[29];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min(14,m_psc_mod.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min((size_t)14,m_psc_mod.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
