+++ src/pluto_main/Table_Device_StartupScript.cpp
@@ -344,7 +344,7 @@ if (is_null[5])
 return "NULL";
 
 char *buf = new char[101];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Parameter.c_str(), (unsigned long) min(50,m_Parameter.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Parameter.c_str(), (unsigned long) min((size_t)50,m_Parameter.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -410,7 +410,7 @@ if (is_null[10])
 return "NULL";
 
 char *buf = new char[29];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min(14,m_psc_mod.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min((size_t)14,m_psc_mod.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
