+++ src/pluto_main/Table_FAQ.cpp
@@ -301,7 +301,7 @@ if (is_null[2])
 return "NULL";
 
 char *buf = new char[201];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Summary.c_str(), (unsigned long) min(100,m_Summary.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Summary.c_str(), (unsigned long) min((size_t)100,m_Summary.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -315,7 +315,7 @@ if (is_null[3])
 return "NULL";
 
 char *buf = new char[131071];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Content.c_str(), (unsigned long) min(65535,m_Content.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Content.c_str(), (unsigned long) min((size_t)65535,m_Content.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -381,7 +381,7 @@ if (is_null[8])
 return "NULL";
 
 char *buf = new char[29];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min(14,m_psc_mod.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min((size_t)14,m_psc_mod.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
