+++ src/pluto_main/Table_Image.cpp
@@ -261,7 +261,7 @@ if (is_null[0])
 return "NULL";
 
 char *buf = new char[201];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_PK_Image.c_str(), (unsigned long) min(100,m_PK_Image.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_PK_Image.c_str(), (unsigned long) min((size_t)100,m_PK_Image.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -275,7 +275,7 @@ if (is_null[1])
 return "NULL";
 
 char *buf = new char[39];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Date.c_str(), (unsigned long) min(19,m_Date.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Date.c_str(), (unsigned long) min((size_t)19,m_Date.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -367,7 +367,7 @@ if (is_null[8])
 return "NULL";
 
 char *buf = new char[29];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min(14,m_psc_mod.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min((size_t)14,m_psc_mod.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
