+++ src/pluto_main/Table_News.cpp
@@ -316,7 +316,7 @@ if (is_null[1])
 return "NULL";
 
 char *buf = new char[21];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Date.c_str(), (unsigned long) min(10,m_Date.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Date.c_str(), (unsigned long) min((size_t)10,m_Date.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -343,7 +343,7 @@ if (is_null[3])
 return "NULL";
 
 char *buf = new char[81];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Title.c_str(), (unsigned long) min(40,m_Title.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Title.c_str(), (unsigned long) min((size_t)40,m_Title.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -357,7 +357,7 @@ if (is_null[4])
 return "NULL";
 
 char *buf = new char[511];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_ShortSummary.c_str(), (unsigned long) min(255,m_ShortSummary.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_ShortSummary.c_str(), (unsigned long) min((size_t)255,m_ShortSummary.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -371,7 +371,7 @@ if (is_null[5])
 return "NULL";
 
 char *buf = new char[5000000];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_FullText.c_str(), (unsigned long) min(16777215,m_FullText.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_FullText.c_str(), (unsigned long) min((size_t)16777215,m_FullText.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -437,7 +437,7 @@ if (is_null[10])
 return "NULL";
 
 char *buf = new char[29];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min(14,m_psc_mod.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min((size_t)14,m_psc_mod.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
